chat

+++ src/services/message.service.js
@@ -1,6 +1,12 @@
-import Message from '../models/messege.model.js'
+import Message from '../models/message.model.js'
+import mongoose from 'mongoose';
+
+const isValidObjectId = (id) => mongoose.Types.ObjectId.isValid(id);
 
 export const createMessage = async (messageData) => {
+  if (!messageData || !messageData.sender || !messageData.content) {
+    throw new Error('Missing required message fields');
+  }
   try {
     const message = new Message(messageData);
     await message.save();
@@ -11,6 +17,9 @@ export const createMessage = async (messageData) => {
 }
 
 export const deleteMessage = async (messageId) => {
+  if (!isValidObjectId(messageId)) {
+    throw new Error('Invalid message ID');
+  }
   try {
     return await Message.findByIdAndDelete(messageId);
   } catch (error) {
@@ -19,6 +28,12 @@ export const deleteMessage = async (messageId) => {
 }
 
 export const updateMessage = async (messageId, content) => {
+  if (!isValidObjectId(messageId)) {
+    throw new Error('Invalid message ID');
+  }
+  if (!content || typeof content !== 'string') {
+    throw new Error('Invalid content');
+  }
   try {
     return await Message.findByIdAndUpdate(messageId, { content }, { new: true });
   } catch (error) {
@@ -27,6 +42,9 @@ export const updateMessage = async (messageId, content) => {
 }
 
 export const getMessage = async (messageId) => {
+  if (!isValidObjectId(messageId)) {
+    throw new Error('Invalid message ID');
+  }
   try {
     return await Message.findById(messageId);
   } catch (error) {
@@ -34,8 +52,10 @@ export const getMessage = async (messageId) => {
   }
 }
 
-
 export const getMessagesByUser = async (userId) => {
+  if (!isValidObjectId(userId)) {
+    throw new Error('Invalid user ID');
+  }
   try {
     return await Message.find({ sender: userId });
   } catch (error) {
